Fix the case where the start/end aren't straight
In all my previous examples the pieces at A or B were
straight east-west / north-south pieces. But in the
puzzle 11 June 2020, the A piece was actually a north-west
angle piece, which broke the solver.

So I had to allow specifying the A/B pieces (assumed to be straight,
if not specified). Also meant initialising the puzzle not assuming
that the initial direction would be EAST.

The Puzzle code is: `14-74253243-75434322-16NE.17SE.26SW.11NW`